mixnode: restructure node stats module
