+++ mixnode/src/node/listener/connection_handler/packet_processing.rs
@@ -1,7 +1,7 @@
 // Copyright 2020 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::node::node_statistics;
+use crate::node::statistics;
 use crypto::asymmetric::encryption;
 use mixnode_common::packet_processor::error::MixProcessingError;
 pub use mixnode_common::packet_processor::processor::MixProcessingResult;
@@ -15,13 +15,13 @@ pub struct PacketProcessor {
     inner_processor: SphinxPacketProcessor,
 
     /// Responsible for updating metrics data
-    node_stats_update_sender: node_statistics::UpdateSender,
+    node_stats_update_sender: statistics::PacketEventReporter,
 }
 
 impl PacketProcessor {
     pub(crate) fn new(
         encryption_key: &encryption::PrivateKey,
-        node_stats_update_sender: node_statistics::UpdateSender,
+        node_stats_update_sender: statistics::PacketEventReporter,
     ) -> Self {
         PacketProcessor {
             inner_processor: SphinxPacketProcessor::new(encryption_key.into()),
